Rewrite server logic to show "slow" calculations
Before this commit, serviceB (the server) did
not use any function/method to execute some logic.
I.e. when looking at lttng traces, was not easy
to see what was "slow" in the server.

After this commit, we added some "fake" calculation
helper methods, so that it will be easier to look
in lttng traces and get an idea of which function
is slowing down the server.